refactor(layout): Navbar sous 300 lignes — éclatement complet (H12)
- useNavbar.js: état/computed/handlers (titre, fil d'Ariane, menus, logout)
- NavbarBreadcrumbs/NavbarNotifications/NavbarUserMenu: sections présentationnelles
  + CSS déplacé verbatim (avec @media)
- Navbar.vue: orchestrateur 90 lignes (669 → 90)
- Changement interne sans effet observable: clic-extérieur du menu utilisateur
  via event.target.closest('.user-menu') (équivalent au template ref d'origine)
- Tests: useNavbar (logique) + montage des 3 sous-composants; Navbar.test conservé
